Project Alteration
Based on advice given by Ciprian I set the background image of my div to
gears.png. I used the size of the image I wanted to isolate for the
width and height values. I then found the correct position on the
background for that image. The big wheel is correct now, needs to be
done for the rest of the images.